third_party/WebKit/LayoutTests/media/encrypted-media/encrypted-media-playback-multiple-sessions-expected.txt

Test playback of encrypted media using Clear Key key system with multiple sessions.

EVENT(needkey)
EVENT(needkey)
EVENT(message)
EVENT(message)
EVENT(playing)
This is a testharness.js-based test.
PASS Playback using Clear Key key system with multiple sessions. 
Harness: the test ran to completion.
